Support osx-arm64 apple (#176)
* Add enum item for osx-arm64

* Don't let dedup look at folders for filehashing

* Updated dynamic dll loading to support arm64 architecture for OSX

* Added tests for -ColumnMap on Invoke-SqlBulkCopy

* Refactored Hashtable into Dictionary(Of String, String)

* refactoring -ColumnMap handling

---------

+++ source/source.build.ps1
@@ -10,7 +10,16 @@ $Script:envList = @(
     [PSCustomObject]@{framework="net6.0"; env="win-x64"; output="output\bin\PS7\win-x64"}
     [PSCustomObject]@{framework="net6.0"; env="linux-x64"; output="output\bin\PS7\linux-x64"}
     [PSCustomObject]@{framework="net6.0"; env="osx-x64"; output="output\bin\PS7\osx-x64"}
+    [PSCustomObject]@{framework="net6.0"; env="osx-arm64";   output="output\bin\PS7\osx-arm64"}
 )
+if($DebugOnly){
+    $Script:envList = $Script:envList | Where-Object env -eq "win-x64"
+    if($PSEdition -eq "core") {
+        if($DebugOnly) { $Script:envList = $Script:envList | Where-Object framework -eq "net6.0" }
+        else { $Script:envList = $Script:envList | Where-Object framework -eq "net47" }
+    }
+}
+
 
 function dedup([string]$Path) {
     Write-Host "  DeDuplicate: $Path"
@@ -23,7 +32,7 @@ function dedup([string]$Path) {
         $safeFiles = @{}
         $first = $list[0]
         Write-Host "  --$($first | Split-Path -Leaf)" -NoNewline
-        Get-ChildItem -Path $first |
+        Get-ChildItem -Path $first -File |
             Where-Object Name -ne "System.Data.SQLite.dll" | #always exclude this SQLite dll because of native interop binding
             Where-Object Name -ne "SQLite.Interop.dll" | #always exclude this SQLite dll because of native interop binding
             Get-FileHash |
@@ -73,9 +82,13 @@ task Build {
     exec { dotnet publish "SimplySql.Cmdlets" -c $configuration -o "output\bin" -p:Version=$Version -p:AssemblyVersion=$version} | HV "Building SimplySql.Cmdlets ($version)" "."
     
     Move-Item "output\bin\SimplySql.Cmdlets.*" -Destination "output"
-    Remove-Item "output\bin\*" #-Exclude "SimplySql.*" -Recurse
+    Remove-Item "output\bin\*" -Recurse #-Exclude "SimplySql.*" -Recurse
+    if($DebugOnly) {
+        #create folders...
+        New-Item -ItemType Directory -Name ".\output\bin\PS5\" | Out-Null
+        New-Item -ItemType Directory -Name ".\output\bin\PS7\" | Out-Null
+    }
 
-    if($DebugOnly) { $Script:envList = $Script:envList | Where-Object env -eq "win-x64" }
     foreach($env in $Script:envList) {
         #exec { dotnet publish "SimplySql.Cmdlets" -c $Configuration -r $env -o "output\bin\$env"} | HV "Building PlatformSpecific Dependencies $env" "."
         exec {
@@ -91,18 +104,20 @@ task Build {
     Get-ChildItem -Path .\output\* -Include "*.pdb", "*.json" -Recurse | Remove-Item
 
     #get SQLite Interops...
-    exec { dotnet build "SimplySql.Engine" -c $configuration } | HV "SQLite Interop" "."
-    if (Test-Path ".\output\bin\PS5\win-x86\" -PathType Container) {
-        Copy-Item ".\SimplySql.Engine\bin\$configuration\net47\x86\SQLite.Interop.dll" -Destination ".\output\bin\PS5\win-x86"
+    if(-not $DebugOnly -or $PSEdition -eq "desktop") {
+        exec { dotnet build "SimplySql.Engine" -c $configuration } | HV "SQLite Interop (PS5)" "."
+        if (Test-Path ".\output\bin\PS5\win-x86\" -PathType Container) {
+            Copy-Item ".\SimplySql.Engine\bin\$configuration\net47\x86\SQLite.Interop.dll" -Destination ".\output\bin\PS5\win-x86"
+        }
+        if (Test-Path ".\output\bin\PS5\win-x64\" -PathType Container) {
+            Copy-Item ".\SimplySql.Engine\bin\$configuration\net47\x64\SQLite.Interop.dll" -Destination ".\output\bin\PS5\win-x64"
+        }
+        Get-ChildItem -Path ".\SimplySql.Engine\bin" -Directory | Remove-Item -Recurse
     }
-    if (Test-Path ".\output\bin\PS5\win-x64\" -PathType Container) {
-        Copy-Item ".\SimplySql.Engine\bin\$configuration\net47\x64\SQLite.Interop.dll" -Destination ".\output\bin\PS5\win-x64"
-    }
-    Get-ChildItem -Path ".\SimplySql.Engine\bin" -Directory | Remove-Item -Recurse
 }
 
 Task DeDup {
-    if(-not $SkipDedup) {
+    if(-not $DebugOnly) {
         dedup ".\output\bin\PS5"
         dedup ".\output\bin\PS7"
         dedup ".\output\bin"
